+++ src/main/java/com/soo/routine/entity/IconCategory.java
@@ -12,7 +12,7 @@ public class IconCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "icon_category_id")
-    private Integer id;//카테고리번호 PK
+    private Long id;//카테고리번호 PK
 
     private String categoryName;//카테고리이름
     private Boolean pay;//유료여부
